Add files via upload
This interactive animation game is a tribute to the Ocean Gate explorers, pioneers who pushed the boundaries of what was possible. Their story is an inspiration to us all.
To play the game, press the Space key to turn on the engine. Then, use the arrow keys to explore the ocean. If a shark touches you, the game is over.
This game is dedicated to the families and friends of the Ocean Gate explorers. It is our hope that this game will help to keep their memory alive.